moved Category components into folder
